Main/lightweight_fl/requirements_lightweight.txt
tensorflow-lite==2.12.0
numpy>=1.19.2
cryptography>=3.4.7
flask>=2.0.1
requests>=2.26.0
pillow>=8.3.1
scikit-learn>=0.24.2
